// src/models/customer.ts
export class customer {
  id: number;
  name: string;
  alias: string;
  contactNum1: string;
  contactNum2: string;
  contactNum3: string;
  loanDate: Date;
  financeAmount: number;
  emiAmount: number;
  interestAmount: number;
  fileCharge: number;
  fileAmount: number;
  loadDaysPassed: number;
}